+++ packages/core/session.js
@@ -17,7 +17,7 @@ class Session {
     }
   }
 
-  trackError (event) {
+  _track (event) {
     this[event._handledState.unhandled ? '_unhandled' : '_handled'] += 1
   }
 }